gateway/pty: refuse terminal geometry that would abort the daemon
`SpawnParams`/`ResizeParams` carried `rows`/`cols` as bare `u16` with no
ceiling on the path. `Grid::new`/`Grid::resize` allocate `rows * cols`
cells of 16 bytes in one `vec!`, so 65535 x 65535 asks for ~68 GB — and a
Rust allocation failure calls `handle_alloc_error`, which aborts the
process rather than unwinding. The blast radius is the whole daemon,
including the vault it holds, and it is reachable by a malformed request,
not only a malicious one.

`check_dimensions` is one body called from both faces: a bound enforced
on one of two faces is not a bound, and `pty.resize` is the cheaper way
in because it never consults `[policies.terminal] enabled`. Refused
rather than silently clamped, and the message names the ceiling.

The client's `render::viewport_cells` clamps to `f64::from(u16::MAX)` —
the largest value this server will parse, not one it survives. Left
alone: client-side clamping cannot be the enforcement point.

Also widens `Grid::put`'s wrap check to `u32`. `cursor_col + w` overflows
`u16` once `cols >= 65534`, and a wrapped sum reads as "no wrap needed".
Unreachable under the new ceiling, but the arithmetic must not depend on
a bound enforced in another module.

+++ src/gateway/pty/screen/grid.rs
@@ -204,7 +204,17 @@ impl Grid {
             return;
         }
         let w = width as u16;
-        if self.cursor_col + w > self.cols {
+        // Widened to `u32` before the compare: `cursor_col` may legitimately
+        // sit AT `cols` (the invariant is `cursor_col <= cols` — parking there
+        // means "a wrap is owed"), so on a maximally wide grid `cursor_col + w`
+        // overflows `u16` — panic in debug, wrap in release, and a wrapped sum
+        // is `< cols`, so the wrap check says "no wrap needed" and `idx` then
+        // indexes past the row. Today `handlers::pty::MAX_TERMINAL_DIMENSION`
+        // keeps `cols` three orders of magnitude below that, but the arithmetic
+        // must not depend on a bound enforced in another module — still less on
+        // the older accident that the allocation for such a grid aborted the
+        // process first.
+        if u32::from(self.cursor_col) + u32::from(w) > u32::from(self.cols) {
             self.newline();
             self.cursor_col = 0;
         }
